+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 
 ## [unreleased]
 
+ * Portal: It is now guaranteed that all client-side (browser) errors are actually sent to the server and logged there, 
+   even if there is a huge amount of errors and also on the browser unload event. 
  * Core: Fixed deprecation warnings from the spdy module. It is now loaded lazy, but since it is no longer maintained 
    and only works with Node.js <= 14 HTTP/2 support is de-facto broken. 
    Shouldn't be a big deal because usual a reverse proxy provides HTTP/2 and TLS.